Add files via upload
In a run, we can see the difference in run time for vertorized version operation and non-vectorized version operation using for loop.

The two approaches achieved the same accuracy. However, the vectorization approach cost 2.00 seconds while the for loop cost 664 seconds. The vectorization approach is 332 times faster than the for loop approach. Which changes as per given number of iterations. 

249958.53830030668
Vectorized Version:2.00009346008ms
249958.5383003188
for loop /Non-Vectorized Version:664.000034332ms